Fixed issue with converting inferred states

+++ src/Bonsai.ML.HiddenMarkovModels/Observations/GaussianObservationsStatistics.cs
@@ -49,7 +49,7 @@ namespace Bonsai.ML.HiddenMarkovModels.Observations
         /// </summary>
         [Description("The sequence of inferred most probable states.")]
         [XmlIgnore]
-        public int[] InferredMostProbableStates { get; set; }
+        public long[] InferredMostProbableStates { get; set; }
 
         /// <summary>
         /// Transforms an observable sequence of <see cref="PyObject"/> into an observable sequence 
@@ -64,7 +64,7 @@ namespace Bonsai.ML.HiddenMarkovModels.Observations
                 var covarianceMatricesPyObj = (double[,,])observationsPyObj.GetArrayAttr("Sigmas");
                 var stdDevsPyObj = DiagonalSqrt(covarianceMatricesPyObj);
                 var batchObservationsPyObj = (double[,])pyObject.GetArrayAttr("batch_observations");
-                var inferredMostProbableStatesPyObj = (int[])pyObject.GetArrayAttr("inferred_most_probable_states");
+                var inferredMostProbableStatesPyObj = (long[])pyObject.GetArrayAttr("inferred_most_probable_states");
 
                 return new GaussianObservationsStatistics
                 {
